refactor: 최적화(useCallback, React.memo)
