+++ src/components/NewsStand/Content/MyNews/NewsStandList/SubscribingCompany/SubscribingCompany.jsx
@@ -1,9 +1,9 @@
-import React from 'react';
+import React, { useCallback } from 'react';
 import styled, { css } from 'styled-components';
 
-const SubscribingCompany = ({ id, company, myTargetNews, setMyTargetNews }) => {
-  const isContentNewsCompany = id === myTargetNews;
-  const changeCurrentNews = ({ target: { id } }) => setMyTargetNews(id);
+const SubscribingCompany = ({ id, company, currentNews, setCurrentNews }) => {
+  const isContentNewsCompany = id === currentNews;
+  const changeCurrentNews = useCallback(({ target: { id } }) => setCurrentNews(id), [id, currentNews]);
   return (
     <NewsPublisher id={id} isContentNewsCompany={isContentNewsCompany} onClick={changeCurrentNews}>
       {company}
@@ -11,7 +11,9 @@ const SubscribingCompany = ({ id, company, myTargetNews, setMyTargetNews }) => {
   );
 };
 
-export default SubscribingCompany;
+export default React.memo(SubscribingCompany, (prev, next) => {
+  return prev.currentNews === next.currentNews;
+});
 
 const NewsPublisher = styled.li`
   box-sizing: border-box;
