+++ src/components/NewsStand/Content/NewsCompany/CompanyInfoCard/CompanyInfoCard.jsx
@@ -1,41 +1,7 @@
 import React from 'react';
-import { useRecoilValue } from 'recoil';
-import { mySubscribeNewsCompanyList } from '@recoilStore/news';
 import { Card, CardBtn } from './CompanyInfoCard.style';
 import { COMPANY_CARD } from '@utils/constant';
-import { useSubscribe } from '@hooks/useSubscribe';
-
-const SubScribeBtn = ({ id }) => {
-  const myNewsStandList = useRecoilValue(mySubscribeNewsCompanyList);
-  const isSubscribeCompany = myNewsStandList.some((subscribeId) => subscribeId === id);
-  const { onSubscribe, onUnSubscribe } = useSubscribe();
-
-  const handleSubscribe = ({
-    target: {
-      dataset: { id },
-    },
-  }) => onSubscribe(id);
-
-  const handleUnSubscribe = ({
-    target: {
-      dataset: { id },
-    },
-  }) => onUnSubscribe(id);
-
-  if (isSubscribeCompany) {
-    return (
-      <button data-id={id} onClick={handleUnSubscribe}>
-        {COMPANY_CARD.UNSUBSCRIPTION}
-      </button>
-    );
-  }
-
-  return (
-    <button data-id={id} onClick={handleSubscribe}>
-      {COMPANY_CARD.SUBSCRIPTION}
-    </button>
-  );
-};
+import SubScribeBtn from './SubScribeBtn';
 
 const CompanyInfoCard = ({ id, company, logoImgUrl }) => {
   return (
@@ -49,4 +15,6 @@ const CompanyInfoCard = ({ id, company, logoImgUrl }) => {
   );
 };
 
-export default CompanyInfoCard;
+export default React.memo(CompanyInfoCard, (prev, next) => {
+  return prev.id === next.id;
+});
